+++ docs/release_notes.rst
@@ -15,10 +15,23 @@ Major Updates
 New Features
 
 * Add flag to disable `# noqa` comment processing in API (#485).
+* Methods, Functions and Nested functions that have a docstring now throw D418 (#511).
+* Methods decorated with @overload no longer reported as D102 (#511).
+* Functions and nested functions decorated with @overload no longer reported as D103 (#511).
 
 Bug Fixes
 
 * Treat "package" as an imperative verb for D401 (#356).
+
+5.1.2 - September 13th, 2020
+----------------------------
+
+New Features
+
+* Methods, Functions and Nested functions that have a docstring now throw D418 (#511).
+* Methods decorated with @overload no longer reported as D102.
+* Functions and nested functions decorated with @overload no longer reported as D103.
+
 
 5.1.1 - August 29th, 2020
 ---------------------------